Replaced iPhone 14 website link with iPhone 16 website link
Apple recently modified the iPhone 14 website and removed grids from the site. As this site was listed in the grid section, I have replaced it with the (similar) iPhone 16 site that does feature grids. The iPhone 16's site does contain less grids, and those that it does have are smaller, so thy may not be as good of an example. However, they heavily resemble those on the old iPhone 14 site, so I feel the substitution is justified. Also, the preview image has not yet been updated. I will update that soon.
I started this as an issue viewable here:
https://github.com/juxtopposed/stacksorted/issues/60